refactor: 결제 관련 route path 구조 수정
