--- src/pages/funding/[id].tsx
@@ -1,5 +1,0 @@
-import Link from 'next/link';
-
-export default function Funding() {
-  return <Link href={{ pathname: '/payment', query: { amount: 20000 } }}>결제하기</Link>;
-}
